Megred renderer

+++ our_gl.h
@@ -14,7 +14,7 @@ void lookat(Vec3f eye, Vec3f center, Vec3f up);
 
 struct IShader {
     virtual ~IShader() {}
-    virtual Vec3i vertex(int iface, int nthvert) = 0;
+    virtual Vec3f vertex(int iface, int nthvert) = 0;
     virtual bool fragment(Vec3f bar, TGAColor &color) = 0;
 };
